[tests] improv: pull out test rules and views to the new module
