+++ moviper-test/src/main/AndroidManifest.xml
@@ -1,0 +1,5 @@
+<manifest package="com.mateuszkoslacz.moviper.tests">
+
+    <application/>
+
+</manifest>
