+++ moviper/build.gradle
@@ -37,6 +37,9 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
+    androidTestCompile('com.android.support.test.espresso:espresso-core:' + rootProject.ext.espressoVersion, {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
     testCompile 'org.mockito:mockito-core:' + rootProject.ext.mockitoVersion
     compile 'com.android.support:appcompat-v7:' + rootProject.ext.supportVersion
@@ -47,6 +50,7 @@ dependencies {
     // TODO: probably should split to two flavours to avoid
     // including rxJava in projects that don't use rx
     compile 'io.reactivex:rxjava:' + rootProject.ext.rxJavaVersion
+    compile 'io.reactivex:rxandroid:' + rootProject.ext.rxAndroidVersion
 
     compile 'com.jakewharton:butterknife:' + rootProject.ext.butterKnifeVersion
     apt 'com.jakewharton:butterknife-compiler:' + rootProject.ext.butterKnifeVersion
