+++ sample-ipc/build.gradle
@@ -42,9 +42,11 @@ dependencies {
     compile 'io.reactivex:rxandroid:' + rootProject.ext.rxAndroidVersion
     compile project(path: ':moviper')
 
+
     androidTestCompile('com.android.support.test.espresso:espresso-core:' + rootProject.ext.espressoVersion, {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+    androidTestCompile project(':moviper-test')
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
     testCompile 'org.mockito:mockito-core:' + rootProject.ext.mockitoVersion
     testCompile 'org.robolectric:robolectric:' + rootProject.ext.robolectricVersion
