+++ sample-ipc/src/debug/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application>
-        <activity android:name="com.mateuszkoslacz.moviper.ipcsample.TestActivity"
+        <activity android:name="com.mateuszkoslacz.moviper.tests.TestActivity"
                 android:theme="@style/Theme.AppCompat"/>
     </application>
 </manifest>
