+++ sample-recyclerview/src/debug/AndroidManifest.xml
@@ -2,9 +2,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application>
-        <activity android:name="com.mateuszkoslacz.moviper.recyclerviewsample.utils.ViewHolderInstrumentedTestActivity"
+        <activity android:name="com.mateuszkoslacz.moviper.tests.ViewHolderInstrumentedTestActivity"
                   android:theme="@style/Theme.AppCompat"/>
-        <activity android:name="com.mateuszkoslacz.moviper.recyclerviewsample.utils.ViewHolderUnitTestActivity"
+        <activity android:name="com.mateuszkoslacz.moviper.tests.ViewHolderUnitTestActivity"
                   android:theme="@style/Theme.AppCompat"/>
     </application>
 </manifest>
