+++ sample-rx-ai/src/test/java/com/mateuszkoslacz/moviper/rxsample/viper/presenter/ListingPresenterTest.java
@@ -2,9 +2,9 @@ package com.mateuszkoslacz.moviper.rxsample.viper.presenter;
 
 import com.mateuszkoslacz.moviper.rxsample.viper.entity.User;
 import com.mateuszkoslacz.moviper.rxsample.viper.interactor.ListingInteractor;
-import com.mateuszkoslacz.moviper.rxsample.viper.utils.RxAndroidSchedulersOverrideRule;
 import com.mateuszkoslacz.moviper.rxsample.viper.routing.ListingRouting;
 import com.mateuszkoslacz.moviper.rxsample.viper.view.activity.ListingActivity;
+import com.mateuszkoslacz.moviper.tests.rules.RxAndroidSchedulersOverrideRule;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -12,7 +12,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -21,7 +21,10 @@ import java.util.List;
 import rx.schedulers.TestScheduler;
 import rx.subjects.TestSubject;
 
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 /**
  * Created by mateuszkoslacz on 17.11.2016.
