+++ sample-rx-viewstate/build.gradle
@@ -53,4 +53,5 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
+    testCompile project(':moviper-test')
 }
