+++ sample-rx/build.gradle
@@ -56,4 +56,5 @@ dependencies {
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
     testCompile 'org.mockito:mockito-core:' + rootProject.ext.mockitoVersion
     testCompile 'org.robolectric:robolectric:' + rootProject.ext.robolectricVersion
+    testCompile project(':moviper-test')
 }
