+++ sample-super-rx-ai-kotlin/build.gradle
@@ -51,6 +51,7 @@ dependencies {
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
     testCompile 'org.mockito:mockito-core:' + rootProject.ext.mockitoVersion
     testCompile 'org.robolectric:robolectric:' + rootProject.ext.robolectricVersion
+    testCompile project(':moviper-test')
 
     testCompile 'org.khronos:opengl-api:' + rootProject.ext.openglApiVersion
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
