+++ sample-super-rx-databinding/build.gradle
@@ -59,4 +59,5 @@ dependencies {
     testCompile 'junit:junit:' + rootProject.ext.junitVersion
     testCompile 'org.mockito:mockito-core:' + rootProject.ext.mockitoVersion
     testCompile 'org.robolectric:robolectric:' + rootProject.ext.robolectricVersion
+    testCompile project(':moviper-test')
 }
