+++ sample-super-rx-databinding/src/test/java/com/mateuszkoslacz/moviper/rxsample/viper/presenter/ListingPresenterWithoutViewTest.java
@@ -5,7 +5,7 @@ import com.mateuszkoslacz.moviper.rxsample.viper.entity.User;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.IOException;
 import java.util.ArrayList;
